requirements.txt
youtube-transcript-api
anthropic
ebooklib
assemblyai
python-dotenv
python-telegram-bot
dropbox
